renamed walker file to lowering
